#280: replacing HTTP backend, 'HttpURLConnectionBackend' doesn't support PATCH; also updating config in our agent tests, also moving envelope related DTOs to model/ package
