+++ agent/src/test/resources/reference.conf
@@ -11,4 +11,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-atum.dispatcher.type="console"
+# dispatcher to be used (http, console, capture)
+atum.dispatcher.type="http"
+
+# The REST API URI of the atum server
+atum.dispatcher.http.url="http://localhost:8080"
+
+# Maximum number of dispatch captures to keep in memory
+#atum.dispatcher.capture.capture-limit=1000 # 0 means no limit
+
