+++ model/src/main/scala/za/co/absa/atum/model/envelopes/Pagination.scala
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package za.co.absa.atum.server.model
+package za.co.absa.atum.model.envelopes
 
-import io.circe.{Decoder, Encoder}
 import io.circe.generic.semiauto.{deriveDecoder, deriveEncoder}
+import io.circe.{Decoder, Encoder}
 
 case class Pagination(
   limit: Int,
