+++ project/Dependencies.scala
@@ -38,7 +38,9 @@ object Dependencies {
 
     val sparkCommons = "0.6.1"
 
-    val sttp = "3.5.2"
+    val sttpClient = "3.5.2"
+    val sttpOkHttpBackend = "1.7.2"
+    val sttpCirceJson = "3.9.7"
 
     val postgresql = "42.6.0"
 
@@ -56,7 +58,6 @@ object Dependencies {
     val http4sBlazeBackend = "0.23.15"
     val http4sPrometheus = "0.23.6"
     val circeJson = "0.14.7"
-    val sttpCirceJson = "3.9.7"
 
     val awssdk = "2.23.15"
 
@@ -83,7 +84,7 @@ object Dependencies {
 
     Seq(
       scalatest,
-      mockito,
+      mockito
     )
   }
 
@@ -97,7 +98,7 @@ object Dependencies {
     Seq(
       circeCore,
       circeParser,
-      circeGeneric,
+      circeGeneric
     )
   }
 
@@ -198,7 +199,8 @@ object Dependencies {
     lazy val sparkCommons = "za.co.absa" % s"spark-commons-spark${sparkMinorVersion}_$scalaMinorVersion" % Versions.sparkCommons
     lazy val sparkCommonsTest = "za.co.absa" % s"spark-commons-test_$scalaMinorVersion" % Versions.sparkCommons % Test
 
-    lazy val sttp = "com.softwaremill.sttp.client3" %% "core" % Versions.sttp
+    lazy val sttpClient3 = "com.softwaremill.sttp.client3" %% "core" % Versions.sttpClient
+    lazy val sttpOkHttpBackend = "com.softwaremill.sttp.client3" %% "okhttp-backend" % Versions.sttpClient
 
     lazy val logback = "ch.qos.logback" % "logback-classic" % Versions.logback
 
@@ -210,7 +212,8 @@ object Dependencies {
       typeSafeConfig,
       sparkCommons,
       sparkCommonsTest,
-      sttp,
+      sttpClient3,
+      sttpOkHttpBackend,
       logback,
       nameOf
     ) ++
@@ -245,7 +248,7 @@ object Dependencies {
       scalaTest,
       balta,
       circe,
-      parser,
+      parser
     )
   }
