+++ server/src/main/scala/za/co/absa/atum/server/api/controller/BaseController.scala
@@ -17,11 +17,12 @@
 package za.co.absa.atum.server.api.controller
 
 import io.circe.{Decoder, parser}
+import za.co.absa.atum.model.envelopes.{ConflictErrorResponse, ErrorInDataErrorResponse, ErrorResponse, InternalServerErrorResponse, NotFoundErrorResponse, PaginatedResult, Pagination}
 import za.co.absa.atum.server.api.exception.ServiceError
 import za.co.absa.atum.server.api.exception.ServiceError._
 import za.co.absa.atum.server.api.http.ApiPaths
-import za.co.absa.atum.server.model.PaginatedResult.{ResultHasMore, ResultNoMore}
-import za.co.absa.atum.server.model.SuccessResponse._
+import za.co.absa.atum.model.envelopes.PaginatedResult.{ResultHasMore, ResultNoMore}
+import za.co.absa.atum.model.envelopes.SuccessResponse._
 import za.co.absa.atum.server.model._
 import zio._
