+++ server/src/main/scala/za/co/absa/atum/server/api/controller/FlowControllerImpl.scala
@@ -17,9 +17,9 @@
 package za.co.absa.atum.server.api.controller
 
 import za.co.absa.atum.model.dto.{CheckpointDTO, CheckpointQueryDTO}
+import za.co.absa.atum.model.envelopes.ErrorResponse
 import za.co.absa.atum.server.api.service.FlowService
-import za.co.absa.atum.server.model.ErrorResponse
-import za.co.absa.atum.server.model.SuccessResponse.MultiSuccessResponse
+import za.co.absa.atum.model.envelopes.SuccessResponse.MultiSuccessResponse
 import zio._
 
 class FlowControllerImpl(flowService: FlowService) extends FlowController with BaseController {
