+++ server/src/main/scala/za/co/absa/atum/server/api/controller/PartitioningControllerImpl.scala
@@ -17,11 +17,11 @@
 package za.co.absa.atum.server.api.controller
 
 import za.co.absa.atum.model.dto._
+import za.co.absa.atum.model.envelopes.{ErrorResponse, GeneralErrorResponse, InternalServerErrorResponse, PaginatedResult}
 import za.co.absa.atum.server.api.exception.ServiceError
 import za.co.absa.atum.server.api.http.ApiPaths.V2Paths
 import za.co.absa.atum.server.api.service.PartitioningService
-import za.co.absa.atum.server.model.SuccessResponse._
-import za.co.absa.atum.server.model.{ErrorResponse, GeneralErrorResponse, InternalServerErrorResponse, PaginatedResult}
+import za.co.absa.atum.model.envelopes.SuccessResponse._
 import zio._
 
 class PartitioningControllerImpl(partitioningService: PartitioningService)
