+++ server/src/main/scala/za/co/absa/atum/server/api/http/BaseEndpoints.scala
@@ -23,6 +23,7 @@ import za.co.absa.atum.server.model._
 import sttp.tapir.typelevel.MatchType
 import sttp.tapir.ztapir._
 import sttp.tapir.{EndpointOutput, PublicEndpoint}
+import za.co.absa.atum.model.envelopes.{BadRequestResponse, ConflictErrorResponse, ErrorInDataErrorResponse, ErrorResponse, GeneralErrorResponse, InternalServerErrorResponse, NotFoundErrorResponse}
 import za.co.absa.atum.server.api.http.ApiPaths._
 
 import java.util.UUID
