+++ server/src/main/scala/za/co/absa/atum/server/api/http/ServerOptions.scala
@@ -27,7 +27,7 @@ import sttp.tapir.server.interceptor.decodefailure.DefaultDecodeFailureHandler.r
 import sttp.tapir.server.interceptor.metrics.MetricsRequestInterceptor
 import sttp.tapir.server.model.ValuedEndpointOutput
 import sttp.tapir.ztapir.{headers, statusCode}
-import za.co.absa.atum.server.model.BadRequestResponse
+import za.co.absa.atum.model.envelopes.BadRequestResponse
 import zio.interop.catz._
 
 trait ServerOptions {
