+++ server/src/main/scala/za/co/absa/atum/server/api/repository/CheckpointRepository.scala
@@ -17,8 +17,8 @@
 package za.co.absa.atum.server.api.repository
 
 import za.co.absa.atum.model.dto.{CheckpointDTO, CheckpointV2DTO}
+import za.co.absa.atum.model.envelopes.PaginatedResult
 import za.co.absa.atum.server.api.exception.DatabaseError
-import za.co.absa.atum.server.model.PaginatedResult
 import zio._
 import zio.macros.accessible
