+++ server/src/main/scala/za/co/absa/atum/server/api/service/CheckpointService.scala
@@ -17,8 +17,8 @@
 package za.co.absa.atum.server.api.service
 
 import za.co.absa.atum.model.dto.{CheckpointDTO, CheckpointV2DTO}
+import za.co.absa.atum.model.envelopes.PaginatedResult
 import za.co.absa.atum.server.api.exception.ServiceError
-import za.co.absa.atum.server.model.PaginatedResult
 import zio.IO
 import zio.macros.accessible
