+++ server/src/main/scala/za/co/absa/atum/server/api/service/CheckpointServiceImpl.scala
@@ -17,9 +17,9 @@
 package za.co.absa.atum.server.api.service
 
 import za.co.absa.atum.model.dto.{CheckpointDTO, CheckpointV2DTO}
+import za.co.absa.atum.model.envelopes.PaginatedResult
 import za.co.absa.atum.server.api.exception.ServiceError
 import za.co.absa.atum.server.api.repository.CheckpointRepository
-import za.co.absa.atum.server.model.PaginatedResult
 import zio._
 
 import java.util.UUID
