+++ server/src/main/scala/za/co/absa/atum/server/api/service/PartitioningServiceImpl.scala
@@ -17,9 +17,9 @@
 package za.co.absa.atum.server.api.service
 
 import za.co.absa.atum.model.dto._
+import za.co.absa.atum.model.envelopes.PaginatedResult
 import za.co.absa.atum.server.api.exception.ServiceError
 import za.co.absa.atum.server.api.repository.PartitioningRepository
-import za.co.absa.atum.server.model.PaginatedResult
 import zio._
 
 class PartitioningServiceImpl(partitioningRepository: PartitioningRepository)
